Working refactored version, except for nontrivial propagators and outputting.
Calculated Config/pac.new with refactored version, gives essentially identical result with original code. (<1-5% difference in peaks and nearly identical elsewhere, probably due to differences in the ordering of calculations within each timestep.)

+++ branches/refactor/tau.h
@@ -16,7 +16,7 @@ class Tau : public NF
 {
   Tau(Tau& ); // no copy constructor
   int max; // if tau is nonhomogeneous, == biggest element
-  vector<double> m; // tau values across nodes, size()==1 if homogeneous
+  vector<int> m; // tau values across nodes, size()==1 if homogeneous
 protected:
   void init( Configf& configf );
   void restart( Restartf& restartf );
